그래픽 업데이트 : Ultimate asset에 있는 백그라운드, Details 이미지로 바꿈 Map 36 제작

+++ Assets/Scripts/GameManager.cs
@@ -38,7 +38,7 @@ public class GameManager : MonoBehaviour, IPointerClickHandler
     [SerializeField] private GameObject Player;
 
     public static float Health = 100;
-    public static int RecentHeal = 0;
+    public static float RecentHeal = 0;
     public static int RecentVirus = 0;
 
     private float time;
